+++ config/reader/options.go
@@ -2,11 +2,8 @@ package reader
 
 import (
 	"c-z.dev/go-micro/config/encoder"
-	"c-z.dev/go-micro/config/encoder/hcl"
 	"c-z.dev/go-micro/config/encoder/json"
-	"c-z.dev/go-micro/config/encoder/toml"
 	"c-z.dev/go-micro/config/encoder/xml"
-	"c-z.dev/go-micro/config/encoder/yaml"
 )
 
 type Options struct {
@@ -19,11 +16,7 @@ func NewOptions(opts ...Option) Options {
 	options := Options{
 		Encoding: map[string]encoder.Encoder{
 			"json": json.NewEncoder(),
-			"yaml": yaml.NewEncoder(),
-			"toml": toml.NewEncoder(),
 			"xml":  xml.NewEncoder(),
-			"hcl":  hcl.NewEncoder(),
-			"yml":  yaml.NewEncoder(),
 		},
 	}
 	for _, o := range opts {
