+++ config/source/etcd/etcd.go
@@ -39,7 +39,7 @@ func (c *etcd) Read() (*source.ChangeSet, error) {
 
 	kvs := make([]*mvccpb.KeyValue, 0, len(rsp.Kvs))
 	for _, v := range rsp.Kvs {
-		kvs = append(kvs, (*mvccpb.KeyValue)(v))
+		kvs = append(kvs, v)
 	}
 
 	data := makeMap(c.opts.Encoder, kvs, c.stripPrefix)
