+++ config/source/etcd/util.go
@@ -37,7 +37,7 @@ func makeMap(e encoder.Encoder, kv []*mvccpb.KeyValue, stripPrefix string) map[s
 }
 
 func update(e encoder.Encoder, data map[string]interface{}, v *mvccpb.KeyValue, action, stripPrefix string) map[string]interface{} {
-	// remove prefix if non empty, and ensure leading / is removed as well
+	// remove prefix if non-empty, and ensure leading / is removed as well
 	vkey := strings.TrimPrefix(strings.TrimPrefix(string(v.Key), stripPrefix), "/")
 	// split on prefix
 	haveSplit := strings.Contains(vkey, "/")
